+++ cs/Markdown/Markdown/Document/Document.cs
@@ -1,0 +1,18 @@
+using System.Collections.Generic;
+
+namespace Markdown
+{
+    public class Document
+    {
+        public TreeNode RootNode { get; }
+
+        public Document()
+        {
+            RootNode = new TreeNode()
+            {
+                IsRoot = true,
+                Child = new List<TreeNode>()
+            };
+        }
+    }
+}
