+++ cs/Markdown/Markdown/Document/NodeType.cs
@@ -1,0 +1,9 @@
+namespace Markdown
+{
+    public enum NodeType
+    {
+        SingleUnderlineTag,
+        DoubleUnderlineTag,
+        Text,
+    }
+}
