+++ cs/Markdown/Markdown/Document/TextNode.cs
@@ -1,0 +1,13 @@
+namespace Markdown
+{
+    public class TextNode : TreeNode
+    {
+        public string Text { get; }
+
+        public TextNode(string text)
+        {
+            Type = NodeType.Text;
+            Text = text;
+        }
+    }
+}
